requirements.txt
git+ssh://git@github.com/credgenics/credgenics-kafka-library.git@v1.0.6
psycopg2==2.9.7
python-dotenv==0.15.0